Feature/backend

See merge request s03-blockchain-sub3/s03p23a103!129

+++ frontend/public/contract/RatingRecord.sol
@@ -2,24 +2,34 @@ pragma solidity ^0.5.12;
 import "./Ownable.sol";
 
 interface RatingRecordInterface {
-    function addRating(address getter, address purchaseContractAddress, uint score) external;
+    function addRating(
+        address getter,
+        address purchaseContractAddress,
+        uint256 score
+    ) external;
 }
 
 contract RatingRecord is RatingRecordInterface, Ownable {
-
     struct Rating {
         address evaluator;
-        uint score;
+        uint256 score;
     }
 
-//특정 판매자 purchase의 평가
-    mapping (address => mapping(address => Rating)) ratingRecords;
-    mapping (address => uint) ratingCounts;
+    //특정 판매자 purchase의 평가
+    mapping(address => mapping(address => Rating)) ratingRecords;
+    mapping(address => uint256) ratingCounts;
 
-    function addRating(address getter, address purchaseContractAddress, uint score) external {
-        ratingRecords[getter][purchaseContractAddress] = Rating(msg.sender, score);
+    function addRating(
+        address getter,
+        address purchaseContractAddress,
+        uint256 score
+    ) external {
+        ratingRecords[getter][purchaseContractAddress] = Rating(
+            msg.sender,
+            score
+        );
 
-        uint ratingCount = ratingCounts[getter];
-        ratingCounts[getter] = uint(ratingCount + 1);
+        uint256 ratingCount = ratingCounts[getter];
+        ratingCounts[getter] = uint256(ratingCount + 1);
     }
 }
